ring-ping: only use nodes with working dual-stack connectivity
Change ring-ping so that it only operates on nodes with working dual-stack
connectivity (i.e., with `alive_ipv4 == alive_ipv6 == 1`).

This should reduce the amount of false errors caused by connectivity issues
local to the nodes themselves.

Also add a mention in the help output that `-n 0` means «all nodes».